完善SuperService 和 SuperDao的实现与继承问题

+++ src/main/java/cn/tata/t2s/ssm/service/NormalUserService.java
@@ -46,7 +46,7 @@ public interface NormalUserService {
 	// 管理员
 	// public void saveTopicState(int projectId, String state_name);
 
-	public void removePersonTopic(int topicId, String personId);
+	public void removePersonTopic(long topicId, String personId);
 
 	public void savePersonReply(Reply reply);
